Add InvoicesTable and PlansCards components to BillingView
- Introduced the InvoicesTable component for displaying user invoices, enhancing the billing overview with a structured table format.
- Added the PlansCards component to showcase available subscription plans, allowing users to easily view and upgrade their subscriptions.
- Refactored the BillingView component to utilize these new components, improving code modularity and readability.
- Ensured localization support is maintained for billing-related terms, enhancing accessibility for users.
- All changes adhere to coding standards and maintain a focus on clean, modular code.
